@@ -8,7 +8,6 @@
 #include <iostream>
 #include "Demos/Visualization/Visualization.h"
 #include "Simulation/DistanceFieldCollisionDetection.h"
-#include "Utils/OBJLoader.h"
 #include "Utils/Logger.h"
 #include "Utils/Timing.h"
 #include "Utils/FileSystem.h"
@@ -149,54 +148,6 @@ void timeStep ()
 		model->getTetModels()[i]->updateMeshNormals(model->getParticles());
 }
 
-void loadObj(const std::string &filename, VertexData &vd, IndexedFaceMesh &mesh, const Vector3r &scale)
-{
-	std::vector<OBJLoader::Vec3f> x;
-	std::vector<OBJLoader::Vec3f> normals;
-	std::vector<OBJLoader::Vec2f> texCoords;
-	std::vector<MeshFaceIndices> faces;
-	OBJLoader::Vec3f s = { (float)scale[0], (float)scale[1], (float)scale[2] };
-	OBJLoader::loadObj(filename, &x, &faces, &normals, &texCoords, s);
-
-	mesh.release();
-	const unsigned int nPoints = (unsigned int)x.size();
-	const unsigned int nFaces = (unsigned int)faces.size();
-	const unsigned int nTexCoords = (unsigned int)texCoords.size();
-	mesh.initMesh(nPoints, nFaces * 2, nFaces);
-	vd.reserve(nPoints);
-	for (unsigned int i = 0; i < nPoints; i++)
-	{
-		vd.addVertex(Vector3r(x[i][0], x[i][1], x[i][2]));
-	}
-	for (unsigned int i = 0; i < nTexCoords; i++)
-	{
-		mesh.addUV(texCoords[i][0], texCoords[i][1]);
-	}
-	for (unsigned int i = 0; i < nFaces; i++)
-	{
-		int posIndices[3];
-		int texIndices[3];
-		for (int j = 0; j < 3; j++)
-		{
-			posIndices[j] = faces[i].posIndices[j];
-			if (nTexCoords > 0)
-			{
-				texIndices[j] = faces[i].texIndices[j];
-				mesh.addUVIndex(texIndices[j]);
-			}
-		}
-
-		mesh.addFace(&posIndices[0]);
-	}
-	mesh.buildNeighbors();
-
-	mesh.updateNormals(vd, 0);
-	mesh.updateVertexNormals(vd);
-
-	LOG_INFO << "Number of triangles: " << nFaces;
-	LOG_INFO << "Number of vertices: " << nPoints;
-}
-
 void buildModel ()
 {
 	TimeManager::getCurrent ()->setTimeStepSize (static_cast<Real>(0.005));
@@ -208,13 +159,13 @@ void buildModel ()
 	string fileName = FileSystem::normalizePath(base->getExePath() + "/resources/models/cube.obj");
 	IndexedFaceMesh mesh;
 	VertexData vd;
-	loadObj(fileName, vd, mesh, Vector3r::Ones());	
+	DemoBase::loadMesh(fileName, vd, mesh, Vector3r::Zero(), Matrix3r::Identity(), Vector3r::Ones());
 	mesh.setFlatShading(true);
 
 	string fileNameTorus = FileSystem::normalizePath(base->getExePath() + "/resources/models/torus.obj");
 	IndexedFaceMesh meshTorus;
 	VertexData vdTorus;
-	loadObj(fileNameTorus, vdTorus, meshTorus, Vector3r::Ones());
+	DemoBase::loadMesh(fileNameTorus, vdTorus, meshTorus, Vector3r::Zero(), Matrix3r::Identity(), Vector3r::Ones());
 
 	SimulationModel::RigidBodyVector &rb = model->getRigidBodies();
 	rb.resize(2);
