@@ -6,7 +6,6 @@
 #include "Simulation/SimulationModel.h"
 #include "Simulation/TimeStepController.h"
 #include <iostream>
-#include "Utils/OBJLoader.h"
 #include "Demos/Visualization/Visualization.h"
 #include "Utils/Logger.h"
 #include "Utils/Timing.h"
@@ -141,54 +140,6 @@ void render ()
 	base->render();
 }
 
-void loadObj(const std::string &filename, VertexData &vd, IndexedFaceMesh &mesh, const Vector3r &scale)
-{
-	std::vector<OBJLoader::Vec3f> x;
-	std::vector<OBJLoader::Vec3f> normals;
-	std::vector<OBJLoader::Vec2f> texCoords;
-	std::vector<MeshFaceIndices> faces;
-	OBJLoader::Vec3f s = { (float)scale[0], (float)scale[1], (float)scale[2] };
-	OBJLoader::loadObj(filename, &x, &faces, &normals, &texCoords, s);
-
-	mesh.release();
-	const unsigned int nPoints = (unsigned int)x.size();
-	const unsigned int nFaces = (unsigned int)faces.size();
-	const unsigned int nTexCoords = (unsigned int)texCoords.size();
-	mesh.initMesh(nPoints, nFaces * 2, nFaces);
-	vd.reserve(nPoints);
-	for (unsigned int i = 0; i < nPoints; i++)
-	{
-		vd.addVertex(Vector3r(x[i][0], x[i][1], x[i][2]));
-	}
-	for (unsigned int i = 0; i < nTexCoords; i++)
-	{
-		mesh.addUV(texCoords[i][0], texCoords[i][1]);
-	}
-	for (unsigned int i = 0; i < nFaces; i++)
-	{
-		int posIndices[3];
-		int texIndices[3];
-		for (int j = 0; j < 3; j++)
-		{
-			posIndices[j] = faces[i].posIndices[j];
-			if (nTexCoords > 0)
-			{
-				texIndices[j] = faces[i].texIndices[j];
-				mesh.addUVIndex(texIndices[j]);
-			}
-		}
-
-		mesh.addFace(&posIndices[0]);
-	}
-	mesh.buildNeighbors();
-
-	mesh.updateNormals(vd, 0);
-	mesh.updateVertexNormals(vd);
-
-	LOG_INFO << "Number of triangles: " << nFaces;
-	LOG_INFO << "Number of vertices: " << nPoints;
-}
-
 /** Create the bunny rod body model
 */
 void createBunnyRodModel()
@@ -200,13 +151,13 @@ void createBunnyRodModel()
 	string fileName = FileSystem::normalizePath(base->getExePath() + "/resources/models/cube.obj");
 	IndexedFaceMesh mesh;
 	VertexData vd;
-	loadObj(fileName, vd, mesh, Vector3r(height, width, depth));
+	DemoBase::loadMesh(fileName, vd, mesh, Vector3r::Zero(), Matrix3r::Identity(), Vector3r(height, width, depth));
 	mesh.setFlatShading(true);
 
 	string fileName2 = FileSystem::normalizePath(base->getExePath() + "/resources/models/bunny_10k.obj");
 	IndexedFaceMesh mesh2;
 	VertexData vd2;
-	loadObj(fileName2, vd2, mesh2, Vector3r(2.0, 2.0, 2.0)); 
+	DemoBase::loadMesh(fileName2, vd2, mesh2, Vector3r::Zero(), Matrix3r::Identity(), Vector3r(2.0, 2.0, 2.0));
 
 	rb.resize(numberOfBodies);	
 	for (unsigned int i = 0; i < numberOfBodies-1; i++)
