@@ -1,12 +1,12 @@
-# -*- coding: utf-8 -*-
+from typing import Union
 
 
-def _to_int(s):
+def _to_int(s: str) -> Union[int, str]:
     try:
         return int(s)
     except ValueError:
         return s
 
 
-__version__ = "2.6.1.dev1"
+__version__ = "2.10.1"
 version_info = tuple(_to_int(s) for s in __version__.split("."))
